+++ be/src/column/binary_column.h
@@ -196,7 +196,7 @@ public:
     }
 
     void append_default(size_t count) override {
-        _offsets.insert(_offsets.end(), count, _bytes.size());
+        _offsets.insert(_offsets.end(), count, static_cast<uint32_t>(_bytes.size()));
         _slices_cache = false;
     }
 
@@ -221,7 +221,7 @@ public:
 
     uint32_t serialize_size(size_t idx) const override {
         // max size of one string is 2^32, so use sizeof(uint32_t) not sizeof(T)
-        return sizeof(uint32_t) + _offsets[idx + 1] - _offsets[idx];
+        return static_cast<uint32_t>(sizeof(uint32_t) + _offsets[idx + 1] - _offsets[idx]);
     }
 
     MutableColumnPtr clone_empty() const override { return BinaryColumnBase<T>::create_mutable(); }
