+++ be/src/column/object_column.cpp
@@ -228,8 +228,9 @@ void ObjectColumn<T>::fnv_hash(uint32_t* hash, uint32_t from, uint32_t to) const
     std::string s;
     for (uint32_t i = from; i < to; ++i) {
         s.resize(_pool[i].serialize_size());
-        int32_t size = _pool[i].serialize(reinterpret_cast<uint8_t*>(s.data()));
-        hash[i] = HashUtil::fnv_hash(s.data(), size, hash[i]);
+        //TODO: May be overflow here if the object is large then 2G.
+        size_t size = _pool[i].serialize(reinterpret_cast<uint8_t*>(s.data()));
+        hash[i] = HashUtil::fnv_hash(s.data(), static_cast<int32_t>(size), hash[i]);
     }
 }
